added custom objects task
