+++ Task07a_custom_types.js
@@ -1,11 +1,10 @@
 const {createCustomType} = require('./handson/customTypes');
-const {createExtension} = require('./handson/extensions');
 const {log} = require('./logger');
 
 
 
 
-const sampleTypeDraft = {
+const sampleTypeDraftData = {
     key:'allowed-to-place-orders',
     name:{
         "de-DE":'allowed-to-place-orders'
@@ -28,18 +27,4 @@ const sampleTypeDraft = {
 
 }
 
-//createCustomType(sampleTypeDraft).then(log).catch(log);
-
-const sampleExtensionDraft = {
-    key:'orderChecker',
-    destination:{
-        type:'HTTP',
-        url:'https://europe-west3-ct-support.cloudfunctions.net/training-extensions-sample'
-    },
-    triggers:[{
-        resourceTypeId:'order',
-        actions:['Create']
-    }]
-}
-
-createExtension(sampleExtensionDraft).then(log).catch(log);
+createCustomType(sampleTypeDraftData).then(log).catch(log);
